fix(API): finish authorization and basic crud operations over estate and contract entities

+++ HomeHive.WebAPI/Controllers/EstatesController.cs
@@ -1,4 +1,6 @@
-﻿using HomeHive.Application.Contracts.Interfaces;
+﻿using HomeHive.Application.Contracts.Caching;
+using HomeHive.Application.Contracts.Interfaces;
+using HomeHive.Application.Features.Estates.Commands.CreateEstate;
 using HomeHive.Application.Features.Estates.Commands.DeleteEstateById;
 using HomeHive.Application.Features.Estates.Commands.UpdateEstate;
 using HomeHive.Application.Features.Estates.Queries.GetAllEstates;
@@ -10,13 +12,18 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace HomeHive.WebAPI.Controllers;
 
-public class EstatesController(ICurrentUserService currentUserService) : ApiBaseController
-{
+public class EstatesController(ICurrentUserService currentUserService, ITokenCacheService tokenCacheService) : ApiBaseController
+{   
+    [Authorize(Roles = "User, Admin")]
     [HttpPost]
     [ProducesResponseType(StatusCodes.Status201Created)]
     public async Task<IActionResult> Create([FromBody] EstateData data)
     {
-        var result = await Mediator.Send(new CreateEstateCommand(data with {OwnerId = currentUserService.GetCurrentUserId()}));
+        var isTokenRevoked = await tokenCacheService.IsTokenRevokedAsync();
+        if (isTokenRevoked) return Unauthorized(new { message = "Token is revoked" });
+        
+        var ownerId = currentUserService.GetCurrentUserId();
+        var result = await Mediator.Send(new CreateEstateCommand(ownerId, data));
         if (!result.Success)
         {
             return BadRequest(result);
@@ -25,10 +32,14 @@ public class EstatesController(ICurrentUserService currentUserService) : ApiBase
         return CreatedAtAction(nameof(Create), result);
     }
     
+    [Authorize(Roles = "User, Admin")]
     [HttpPut("{estateId}", Name = "Update")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     public async Task<IActionResult> Update([FromRoute] Guid estateId, [FromBody] EstateData data)
     {
+        var isTokenRevoked = await tokenCacheService.IsTokenRevokedAsync();
+        if (isTokenRevoked) return Unauthorized(new { message = "Token is revoked" });
+        
         var result = await Mediator.Send(new UpdateEstateCommand(estateId, data));
         if (!result.Success)
         {
@@ -39,10 +50,13 @@ public class EstatesController(ICurrentUserService currentUserService) : ApiBase
     }
     
     [Authorize(Roles = "User, Admin")]
-    [HttpDelete]
+    [HttpDelete("{estateId}")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     public async Task<IActionResult> Delete(Guid estateId)
     {
+        var isTokenRevoked = await tokenCacheService.IsTokenRevokedAsync();
+        if (isTokenRevoked) return Unauthorized(new { message = "Token is revoked" });
+        
         var result = await Mediator.Send(new DeleteEstateByIdCommand(estateId));
         if (!result.Success)
         {
@@ -57,6 +71,9 @@ public class EstatesController(ICurrentUserService currentUserService) : ApiBase
     [ProducesResponseType(StatusCodes.Status200OK)]
     public async Task<IActionResult> Get(Guid estateId)
     {
+        var isTokenRevoked = await tokenCacheService.IsTokenRevokedAsync();
+        if (isTokenRevoked) return Unauthorized(new { message = "Token is revoked" });
+        
         var result = await Mediator.Send(new GetEstateByIdQuery(estateId));
 
         if (!result.Success)
@@ -73,6 +90,9 @@ public class EstatesController(ICurrentUserService currentUserService) : ApiBase
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<IActionResult> GetAll()
     {
+        var isTokenRevoked = await tokenCacheService.IsTokenRevokedAsync();
+        if (isTokenRevoked) return Unauthorized(new { message = "Token is revoked" });
+        
         var result = await Mediator.Send(new GetAllEstatesQuery());
 
         if (!result.Success)
